Embedded referenced icon font asset `PUIIcon.ttf`
This used to refer to an asset in Registration component, which only correctly resolved
if the Registration component was integrated too in the target application.

Since USR is deprecated in favor of the new PIM component,
integrators should be able to drop the USR dependency entirely.
The reference in CSW prevented that, embedding the .ttf file makes CSW more self contained.